% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imposeMissing.R
\name{attrition}
\alias{attrition}
\title{Generate Attrition Missingness}
\usage{
attrition(
  data,
  prob = NULL,
  timePoints = 1,
  cov = NULL,
  threshold = NULL,
  ignoreCols = NULL
)
}
\arguments{
\item{data}{Dataset.}

\item{prob}{Probability of attrition at each time point.}

\item{timePoints}{Number of time points.}

\item{cov}{Covariate influencing attrition.}

\item{threshold}{Threshold used in attrition mechanism.}

\item{ignoreCols}{Columns excluded from attrition.}
}
\value{
Logical matrix indicating attrition-induced missing values.
}
\description{
Simulates attrition across multiple time points based on specified
probabilities.
}
\keyword{internal}
