% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{changeDupLab}
\alias{changeDupLab}
\title{Handle duplicated parameter labels in lavaan parameter tables}
\usage{
changeDupLab(pt)
}
\description{
Handle duplicated parameter labels in lavaan parameter tables
}
\keyword{internal}
