% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{checkNPD}
\alias{checkNPD}
\title{Check for non–positive-definite covariance matrices}
\usage{
checkNPD(object)
}
\description{
Detects improper solutions where covariance matrices contain
negative eigenvalues.
}
\keyword{internal}
