% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySimResult.R
\name{coef}
\alias{coef}
\alias{coef,SimResult-method}
\title{Extract parameter estimates}
\usage{
coef(object, ...)

\S4method{coef}{SimResult}(object, improper = TRUE, nonconverged = FALSE)
}
\arguments{
\item{object}{A \code{SimResult} object.}

\item{...}{Additional arguments passed to methods.}

\item{improper}{Logical indicating whether improper solutions should be included.}

\item{nonconverged}{Logical indicating whether nonconverged replications should be included.}
}
\value{
A matrix of parameter estimates.
}
\description{
Method for extracting parameter estimates from a \code{SimResult} object.
}
