% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{collapseExtraParam}
\alias{collapseExtraParam}
\title{Collapse extra parameters generated by equality constraints}
\usage{
collapseExtraParam(pls, dgen, fill = TRUE, con = NULL)
}
\description{
Internal helper used to compute derived parameters defined
through lavaan constraints.
}
\keyword{internal}
