\name{combine.object}
\alias{combine.object}
\title{
	Combine two objects (Internal)
}
\description{
	This function is used to combine two objects in the same or similar type together. 
}
\usage{
combine.object(object1, object2, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object1}{
	The first object
}
  \item{object2}{
	The second object
}
  \item{\dots}{
	Additional options
}
}
\details{
	Candidate objects are \code{vector}, \code{matrix}, \code{simMatrix}, \code{simVector}, \code{matrixSet}, and \code{misspecifiedSet}
	
}
\value{
	The combined objects
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)
}
\note{
	Internal Function
}
