\name{combineLoadingExoEndo}
\alias{combineLoadingExoEndo}
\title{
	Combine factor loading from the exogenous and endogenous sides into a single matrix
}
\description{
	Combine factor loading from the exogenous and endogenous sides into a single matrix
}
\usage{
combineLoadingExoEndo(LX, LY, value = 0)
}
\arguments{
  \item{LX}{
	The factor loading matrix in the exogenous side
}
  \item{LY}{
	The factor loading matrix in the endogenous side
}
  \item{value}{
	the values filling in the other values in the resulting matrix
}
}
\value{
	The combined factor loading matrix
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; \email{psunthud@ku.edu})
}
\examples{
# No example
}
