\name{constant.vector}
\alias{constant.vector}
\title{
	Create constant \code{simVector} (Internal)
}
\description{
	Create a constant \code{simVector}
}
\usage{
constant.vector(constant, ni)
}
\arguments{
  \item{constant}{
	Number that is used to be the constant
}
  \item{ni}{
	Number of items
}
}
\value{
	Return constant \code{simVector}
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)
}
\note{
	Internal function
}
\examples{
	#constant.vector(0, 4)
}
