\name{countMACS}
\alias{countMACS}
\title{
	Count the number of elements in the sufficient statistics
}
\description{
	Count the number of elements in the sufficient statistics (mean vector and covariance matrix)
}
\usage{
countMACS(object)
}
\arguments{
  \item{object}{
	The \code{\linkS4class{SimSet}} object
}
}
\value{
	The number of elements in the sufficient statistics
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; \email{psunthud@ku.edu})
}
\examples{
# No example
}
