\name{cov2corMod}
\alias{cov2corMod}
\title{
	Convert a covariance matrix to a correlation matrix
}
\description{
	Convert a covariance matrix to a correlation matrix. This is the \code{cov2cor} function in the base package that takes care of the zero-variance variables
}
\usage{
cov2corMod(V)
}
\arguments{
  \item{V}{
	A covariance matrix
}
}
\value{
	A correlation matrix
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; \email{psunthud@ku.edu})
}
\examples{
# No example
}
