\name{divide.object}
\alias{divide.object}
\title{
	Make a division on each elements of the object (Internal)
}
\description{
	Make a division on each elements of the object 
}
\usage{
divide.object(object, constant, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
	The desired object
}
  \item{constant}{
	Divisor
}
  \item{\dots}{
	Additional options
}
}
\details{
	Candidate objects are \code{vector}, \code{matrix}, and \code{matrixSet}
}
\value{
	The divided objects
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)
}
\note{
	Internal Function
}
