% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{estmodel}
\alias{estmodel}
\alias{estmodel.cfa}
\alias{estmodel.path}
\alias{estmodel.sem}
\title{Specify an Estimation Model for Simulation}
\usage{
estmodel(
  LY = NULL,
  PS = NULL,
  RPS = NULL,
  TE = NULL,
  RTE = NULL,
  BE = NULL,
  VTE = NULL,
  VY = NULL,
  VPS = NULL,
  VE = NULL,
  TY = NULL,
  AL = NULL,
  MY = NULL,
  ME = NULL,
  KA = NULL,
  GA = NULL,
  modelType,
  indLab = NULL,
  facLab = NULL,
  covLab = NULL,
  groupLab = "group",
  ngroups = 1,
  con = NULL
)

estmodel.cfa(
  LY = NULL,
  PS = NULL,
  RPS = NULL,
  TE = NULL,
  RTE = NULL,
  VTE = NULL,
  VY = NULL,
  VPS = NULL,
  VE = NULL,
  TY = NULL,
  AL = NULL,
  MY = NULL,
  ME = NULL,
  KA = NULL,
  GA = NULL,
  indLab = NULL,
  facLab = NULL,
  covLab = NULL,
  groupLab = "group",
  ngroups = 1,
  con = NULL
)

estmodel.path(
  PS = NULL,
  RPS = NULL,
  BE = NULL,
  VPS = NULL,
  VE = NULL,
  AL = NULL,
  ME = NULL,
  KA = NULL,
  GA = NULL,
  indLab = NULL,
  facLab = NULL,
  covLab = NULL,
  groupLab = "group",
  ngroups = 1,
  con = NULL
)

estmodel.sem(
  LY = NULL,
  PS = NULL,
  RPS = NULL,
  TE = NULL,
  RTE = NULL,
  BE = NULL,
  VTE = NULL,
  VY = NULL,
  VPS = NULL,
  VE = NULL,
  TY = NULL,
  AL = NULL,
  MY = NULL,
  ME = NULL,
  KA = NULL,
  GA = NULL,
  indLab = NULL,
  facLab = NULL,
  covLab = NULL,
  groupLab = "group",
  ngroups = 1,
  con = NULL
)
}
\arguments{
\item{LY}{Factor loading matrix.}

\item{PS}{Factor covariance matrix.}

\item{RPS}{Factor correlation matrix.}

\item{TE}{Measurement error covariance matrix.}

\item{RTE}{Measurement error correlation matrix.}

\item{BE}{Structural regression coefficient matrix.}

\item{VTE}{Measurement error variances.}

\item{VY}{Total indicator variances.}

\item{VPS}{Factor variances.}

\item{VE}{Total latent variances.}

\item{TY}{Indicator intercepts.}

\item{AL}{Latent intercepts.}

\item{MY}{Indicator means.}

\item{ME}{Latent means.}

\item{KA}{Regression coefficients from covariates to indicators.}

\item{GA}{Regression coefficients from covariates to latent variables.}

\item{modelType}{Character string specifying the model type:
\code{"cfa"}, \code{"sem"}, or \code{"path"}.}

\item{indLab}{Optional indicator labels.}

\item{facLab}{Optional latent variable labels.}

\item{covLab}{Optional covariate labels.}

\item{groupLab}{Name of the grouping variable.}

\item{ngroups}{Number of groups in the model.}

\item{con}{Optional parameter constraints.}
}
\value{
A \code{SimSem} object representing the estimation model.
}
\description{
Constructs a \code{SimSem} object describing the model used to analyze
simulated data. This function is similar to
\code{\link{model}} but is designed for estimation rather than data
generation.
}
\details{
This function is similar to \code{\link{model}} but is intended for
specifying the analysis model used when fitting simulated datasets.
}
