\name{fillParam}
\alias{fillParam}
\title{
	Fill in other objects based on the parameter values of current objects
}
\description{
	Fill in other objects based on the parameter values of current objects
}
\usage{
fillParam(param, modelType)
}
\arguments{
  \item{param}{
	The \code{linkS4class{MatrixSet}} class that some matrices have not been filled.
}
  \item{modelType}{
	Analysis model type
}
}
\value{
	The same object that all matrices have been filled.
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; \email{psunthud@ku.edu})
}
\examples{
# No example
}
