% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPowerCoverage.R
\name{findCoverage}
\alias{findCoverage}
\title{Find required values for target coverage}
\usage{
findCoverage(coverTable, iv, target)
}
\arguments{
\item{coverTable}{A data frame containing coverage results from a
simulation study.}

\item{iv}{The independent variable whose value should be determined.
This can be specified either by column name or index.}

\item{target}{The desired coverage level.}
}
\value{
A value (or set of values) of the varying parameter that achieves the
desired coverage level.
}
\description{
Computes the value of a varying parameter that yields a specified
coverage level. Coverage is computed as \eqn{1 - power}, and the
function internally calls \code{\link{findPower}} to determine the
corresponding parameter value.
}
\seealso{
\code{\link{findPower}}
}
