% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{findFactorIntercept}
\alias{findFactorIntercept}
\title{Find factor intercepts}
\usage{
findFactorIntercept(beta, factorMean = NULL, gamma = NULL, covmean = NULL)
}
\arguments{
\item{beta}{Regression coefficient matrix among latent variables.}

\item{factorMean}{Vector of factor means.}

\item{gamma}{Optional matrix of regression coefficients from covariates
to latent variables.}

\item{covmean}{Mean vector of covariates.}
}
\value{
A numeric vector containing factor intercepts.
}
\description{
Computes factor intercepts implied by regression coefficients and
factor means.
}
