% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{findFactorMean}
\alias{findFactorMean}
\title{Find factor means}
\usage{
findFactorMean(beta, alpha = NULL, gamma = NULL, covmean = NULL)
}
\arguments{
\item{beta}{Regression coefficient matrix among latent variables.}

\item{alpha}{Factor intercepts.}

\item{gamma}{Optional regression coefficients from covariates.}

\item{covmean}{Mean vector of covariates.}
}
\value{
Numeric vector of latent variable means.
}
\description{
Computes latent variable means implied by regression coefficients
and factor intercepts.
}
