% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{findFactorTotalCov}
\alias{findFactorTotalCov}
\title{Find factor total covariance}
\usage{
findFactorTotalCov(
  beta,
  psi = NULL,
  corPsi = NULL,
  totalVarPsi = NULL,
  errorVarPsi = NULL,
  gamma = NULL,
  covcov = NULL
)
}
\arguments{
\item{beta}{Regression coefficient matrix among latent variables.}

\item{psi}{Residual covariance matrix of latent variables.}

\item{corPsi}{Correlation matrix of latent variables.}

\item{totalVarPsi}{Total variances of latent variables.}

\item{errorVarPsi}{Residual variances of latent variables.}

\item{gamma}{Optional regression coefficients from covariates.}

\item{covcov}{Covariance matrix of covariates.}
}
\value{
Covariance matrix of latent variables.
}
\description{
Computes the total covariance matrix of latent variables implied by
regression coefficients and latent residual covariance structure..
}
