% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{findIndResidualVar}
\alias{findIndResidualVar}
\title{Find indicator residual variances}
\usage{
findIndResidualVar(
  lambda,
  totalFactorCov,
  totalVarTheta = NULL,
  kappa = NULL,
  covcov = NULL
)
}
\arguments{
\item{lambda}{Factor loading matrix.}

\item{totalFactorCov}{Covariance matrix of latent variables.}

\item{totalVarTheta}{Total indicator variances.}

\item{kappa}{Optional regression coefficients from covariates.}

\item{covcov}{Covariance matrix of covariates.}
}
\value{
Numeric vector of residual variances.
}
\description{
Computes residual variances of indicators implied by factor loadings
and total indicator variances.
}
