% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likRatioFit.R
\name{findphist}
\alias{findphist}
\title{Find the Density of a Pair of Values from a 2D Kernel Density Estimate}
\usage{
findphist(value, hist)
}
\arguments{
\item{value}{Numeric vector of length two representing the target pair of values.}

\item{hist}{A 2D binned kernel density estimate produced by
\code{\link[KernSmooth]{bkde2D}}.}
}
\value{
The estimated density corresponding to the pair of values.
}
\description{
Returns the estimated density corresponding to a pair of values from a
two-dimensional kernel density estimate.
}
\keyword{internal}
