% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{findPossibleFactorCor}
\alias{findPossibleFactorCor}
\title{Find possible factor correlations}
\usage{
findPossibleFactorCor(beta)
}
\arguments{
\item{beta}{Regression coefficient matrix among latent variables.}
}
\value{
A matrix indicating which factor correlations may be freely estimated.
}
\description{
Determines which factor correlations may be freely estimated based on
the regression structure among latent variables.
}
