% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPowerCoverage.R
\name{findPower}
\alias{findPower}
\title{Find required values for target power}
\usage{
findPower(powerTable, iv, power)
}
\arguments{
\item{powerTable}{A data frame containing power results from a
simulation study.}

\item{iv}{The independent variable whose value should be determined.
This can be specified either by column name or index.}

\item{power}{The desired level of statistical power.}
}
\value{
A value (or matrix of values) of the varying parameter that achieves
the specified power level.
}
\description{
Determines the value of a varying parameter that achieves a specified
level of statistical power based on a power table from a simulation
study.
}
\details{
The function can handle multiple independent variables by splitting
the data according to the remaining variables and applying the power
search within each subset.
}
\seealso{
\code{\link{findCoverage}}, \code{\link{findTargetPower}}
}
