% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{findRecursiveSet}
\alias{findRecursiveSet}
\title{Find recursive variable sets}
\usage{
findRecursiveSet(beta)
}
\arguments{
\item{beta}{Regression coefficient matrix.}
}
\value{
A list of variable indices representing recursive levels.
}
\description{
Groups variables according to their position in a recursive structural model
defined by the regression coefficient matrix.
}
