% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPowerCoverage.R
\name{findTargetPower}
\alias{findTargetPower}
\title{Find parameter value achieving target power}
\usage{
findTargetPower(iv, dv, power)
}
\arguments{
\item{iv}{A vector of the varying parameter values.}

\item{dv}{A data frame containing power values for the parameters of
interest.}

\item{power}{The desired level of statistical power.}
}
\value{
The value of the varying parameter that provides the desired power.
If \code{NA}, no value in the domain of the varying parameter provides
the desired power. If the minimum value of the varying parameter is
returned, the minimum value already achieves the target power.
}
\description{
Computes the value of a varying parameter that achieves a specified
level of statistical power for a single independent variable.
}
\details{
This function is used internally by \code{\link{findPower}} when only
one varying parameter is present.
}
\keyword{internal}
