% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{fleishman1978_abcd}
\alias{fleishman1978_abcd}
\title{Fleishman polynomial coefficients}
\usage{
fleishman1978_abcd(skewness, kurtosis)
}
\description{
Computes Fleishman (1978) polynomial coefficients for generating
non-normal variables with specified skewness and kurtosis.
}
\keyword{internal}
