% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imposeMissing.R
\name{generateIndices}
\alias{generateIndices}
\title{Generate Sequential Index Groups}
\usage{
generateIndices(ngroups, groupRange, excl = NULL)
}
\arguments{
\item{ngroups}{Number of groups.}

\item{groupRange}{Vector of indices to divide into groups.}

\item{excl}{Optional indices to exclude.}
}
\value{
A list of index vectors.
}
\description{
Creates groups of indices used in planned missing designs.
}
\keyword{internal}
