% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openMx.R
\name{generateMx}
\alias{generateMx}
\title{Generate Data from an OpenMx Model}
\usage{
generateMx(
  object,
  n,
  indDist = NULL,
  groupLab = NULL,
  covData = NULL,
  empirical = FALSE
)
}
\arguments{
\item{object}{OpenMx model object.}

\item{n}{Sample size.}
}
\value{
A data.frame containing simulated data.
}
\description{
Internal utility used by simsem to generate data from an OpenMx model.
}
\keyword{internal}
