% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generateSimSem}
\alias{generateSimSem}
\title{Internal data generation engine for SimSem models}
\usage{
generateSimSem(
  model,
  n,
  maxDraw = 50,
  misfitBounds = NULL,
  misfitType = "f0",
  averageNumMisspec = FALSE,
  optMisfit = NULL,
  optDraws = 50,
  createOrder = c(1, 2, 3),
  indDist = NULL,
  sequential = FALSE,
  facDist = NULL,
  errorDist = NULL,
  saveLatentVar = FALSE,
  indLab = NULL,
  modelBoot = FALSE,
  realData = NULL,
  covData = NULL,
  params = FALSE,
  empirical = FALSE
)
}
\description{
Generates data for a \code{SimSem} model after parameters have been
drawn. This function is used internally by \code{\link{generate}}.
}
\keyword{internal}
