% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySimResult.R
\name{getExtraOutput}
\alias{getExtraOutput}
\title{Get extra outputs from simulation results}
\usage{
getExtraOutput(
  object,
  improper = TRUE,
  nonconverged = FALSE,
  simplify = FALSE,
  USE.NAMES = FALSE
)
}
\arguments{
\item{object}{A \code{\linkS4class{SimResult}} object containing extra
outputs extracted by the function defined in the \code{outfun} argument
of the \code{\link{sim}} function.}

\item{improper}{Logical indicating whether to include replications with
improper solutions.}

\item{nonconverged}{Logical indicating whether to include nonconvergent
replications.}

\item{simplify, USE.NAMES}{See \code{\link{sapply}}.}
}
\value{
A \code{list} of extra outputs, optionally simplified to a vector
or matrix.
}
\description{
Extract extra outputs from a simulation result object
(\code{\linkS4class{SimResult}}). Users can request additional output from
the \code{lavaan} object in each iteration by specifying the \code{outfun}
argument in the \code{\link{sim}} function. See the example below.
}
\examples{
\dontrun{
loading <- matrix(0, 6, 1)
loading[1:6, 1] <- NA
LY <- bind(loading, 0.7)
RPS <- binds(diag(1))
RTE <- binds(diag(6))
CFA.Model <- model(LY = LY, RPS = RPS, RTE = RTE, modelType = "CFA")

# Write a function to extract the modification index from lavaan object
outfun <- function(out) {
  result <- inspect(out, "mi")
}

# Use only 5 replications to save time.
# In reality, more replications are needed.
Output <- sim(5, n = 200, model = CFA.Model, outfun = outfun)

# Get the modification index of each replication
getExtraOutput(Output)
}

}
\seealso{
\itemize{
\item \code{\link{sim}} to run a Monte Carlo simulation.
}
}
