% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySimResult.R
\name{getPopulation}
\alias{getPopulation}
\title{Extract population parameter values}
\usage{
getPopulation(object, std = FALSE, improper = TRUE, nonconverged = FALSE)
}
\arguments{
\item{object}{A \code{SimResult} object.}

\item{std}{Logical indicating whether standardized parameters should be returned.}

\item{improper}{Logical indicating whether improper solutions should be included.}

\item{nonconverged}{Logical indicating whether nonconverged replications should be included.}
}
\value{
Population parameter values.
}
\description{
This function extracts population parameter values from a simulation result object.
}
