% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPowerCoverage.R
\name{getPred}
\alias{getPred}
\title{Construct Predictor Values for Continuous Simulation Conditions}
\usage{
getPred(
  contParam = NULL,
  nVal = NULL,
  pmMCARval = NULL,
  pmMARval = NULL,
  paramVal = NULL
)
}
\arguments{
\item{contParam}{Character vector specifying parameters that vary.}

\item{nVal}{Sample size value.}

\item{pmMCARval}{Proportion of MCAR missingness.}

\item{pmMARval}{Proportion of MAR missingness.}

\item{paramVal}{Values of continuously varying parameters.}
}
\value{
A list containing predictor values.
}
\description{
Creates a list of predictor values used for evaluating power or coverage
when simulation conditions vary continuously.
}
\keyword{internal}
