% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{HeadrickSawilowsky1999}
\alias{HeadrickSawilowsky1999}
\title{Headrick–Sawilowsky nonnormal data generator}
\usage{
HeadrickSawilowsky1999(n = 100L, COR, skewness, kurtosis)
}
\description{
Generates multivariate nonnormal data using the Headrick and
Sawilowsky (1999) method.
}
\keyword{internal}
