% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValue.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate values relative to a baseline vector}
\usage{
interpolate(baselineVec, val, resultVec = NULL)
}
\arguments{
\item{baselineVec}{Baseline vector used for interpolation.}

\item{val}{Value relative to the baseline vector.}

\item{resultVec}{Vector from which resulting values are taken.}
}
\value{
Interpolated value relative to the baseline vector.
}
\description{
Find the value of one vector relative to a value in another vector
using linear interpolation.
}
\details{
If the target value lies between two baseline values, the resulting
value is estimated by linear interpolation.
}
\keyword{internal}
