% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{is.empty}
\alias{is.empty}
\title{Check whether elements are empty}
\usage{
is.empty(dat)
}
\arguments{
\item{dat}{A vector or matrix.}
}
\value{
A logical object of the same shape indicating
whether each element is empty.
}
\description{
Internal helper used to identify empty elements
(either \code{""} or \code{NA}).
}
\keyword{internal}
