% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{is.free}
\alias{is.free}
\title{Identify free parameters}
\usage{
is.free(mat)
}
\arguments{
\item{mat}{A matrix or vector.}
}
\value{
Logical vector indicating free parameters.
}
\description{
Internal helper determining whether parameters are free.
}
\keyword{internal}
