% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{is.label}
\alias{is.label}
\title{Identify character constraint labels}
\usage{
is.label(mat)
}
\arguments{
\item{mat}{A matrix or vector.}
}
\value{
Logical vector indicating whether elements are labels.
}
\description{
Internal helper used to detect equality constraint labels.
}
\keyword{internal}
