\name{is.null.object}
\alias{is.null.object}
\title{
	Check whether the object is NULL (Internal)
}
\description{
	Check whether the object is NULL
}
\usage{
is.null.object(target)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{target}{
	Checked target
}
}
\value{
	TRUE if the object is null. FALSE otherwise.
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)
}
\note{
	Internal Function
}