% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{is.random}
\alias{is.random}
\title{Check to see if random parameters are in the model. Can use to simplfy runRep}
\usage{
is.random(dat)
}
\description{
Check to see if random parameters are in the model. Can use to simplfy runRep
}
\keyword{internal}
