\name{isCorMatrix}
\alias{isCorMatrix}
\title{
	Check whether a \code{matrix} is a possible correlation matrix
}
\description{
	Check whether a \code{matrix} is a possible correlation matrix
}
\usage{
isCorMatrix(matrixA)
}
\arguments{
  \item{matrixA}{
	The target matrix
}
}
\value{
	\code{TRUE} if the target matrix is a possible correlation matrix
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)
}
\examples{
# This function is not a public function.

# isCorMatrix(diag(5))
}
