\name{isDefault}
\alias{isDefault}
\title{
	Check whether a vector object is default
}
\description{
	Check whether a specified the vector object is a default vector object such that users did not specify anything. For example, check whether means of indicators are specified as 1.
}
\usage{
isDefault(object)
}
\arguments{
  \item{object}{
	The target vector object
}
}
\value{
	\code{TRUE} if the target vector object is a default vector object
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)
}
\examples{
# No example
}
