\name{isVarianceConstraint}
\alias{isVarianceConstraint}
\title{
	Check whether all rownames in a constraint matrix containing symbols of variance vectors
}
\description{
	Check whether all rownames in a constraint matrix containing symbols of variance vectors
}
\usage{
isVarianceConstraint(Name)
}
\arguments{
  \item{Name}{
	The rownames of a constraint matrix
}
}
\value{
	\code{TRUE} if all rownames are variances vectors
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)
}
\examples{
# No example
}
