\name{kStat}
\alias{kStat}
\title{
	Calculate the \emph{k}-statistic of a variable
}
\description{
	Calculate the \emph{k}-statistic (i.e., unbiased estimator of a cumulant) of a variable
}
\usage{
kStat(x, ord)
}
\arguments{
  \item{x}{
	A vector of a variable
}
  \item{ord}{
	The order of the \emph{k}-statistics (The maximum value is 4).
}
}
\value{
	The k-statistic value
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)
}
\examples{
# This function is not a public function.

# kStat(1:5, 4)
}
