% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{lavaanValeMaurelli1983}
\alias{lavaanValeMaurelli1983}
\title{Vale–Maurelli nonnormal data generator}
\usage{
lavaanValeMaurelli1983(n = 100L, COR, skewness, kurtosis)
}
\description{
Generates multivariate nonnormal data using the Vale and Maurelli
(1983) method.
}
\keyword{internal}
