% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrnorm.R
\name{lazyCor}
\alias{lazyCor}
\title{Create a Correlation Matrix}
\usage{
lazyCor(X, d)
}
\arguments{
\item{X}{Correlation specification.}

\item{d}{Optional dimension.}
}
\value{
Correlation matrix.
}
\description{
Generates a correlation matrix from flexible inputs:
a scalar correlation, a vech vector, or a full matrix.
}
\author{
Paul Johnson
}
\keyword{internal}
