% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrnorm.R
\name{lazyCov}
\alias{lazyCov}
\title{Create a Covariance Matrix from Correlations and Standard Deviations}
\usage{
lazyCov(Rho, Sd, d)
}
\arguments{
\item{Rho}{Correlation specification.}

\item{Sd}{Standard deviations.}

\item{d}{Optional dimension.}
}
\value{
Covariance matrix.
}
\description{
Constructs a covariance matrix from correlation and standard deviation
information supplied in flexible formats.
}
\author{
Paul Johnson
}
\keyword{internal}
