% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imposeMissing.R
\name{logitMiss}
\alias{logitMiss}
\title{Generate Missingness from Logistic Model}
\usage{
logitMiss(data, script)
}
\arguments{
\item{data}{Dataset.}

\item{script}{Character string specifying logistic missingness syntax.}
}
\value{
Logical matrix indicating missing values.
}
\description{
Applies a logistic regression model to generate missing values
based on user-specified syntax.
}
\keyword{internal}
