\name{make.labels}
\alias{make.labels}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
	Make parameter names for OpenMx (Internal)
}
\description{
	Make parameter names for each element in a matrix or a vector for OpenMx syntax
}
\usage{
make.labels(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{ the target objects
}
  \item{\dots}{
	Name of the desired object, the analysis package, and the attribute told whether the object is symmetric
}
}
\value{
	Return the labels object
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)
}
\note{
	Internal Function
}
