% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imposeMissing.R
\name{makeMAR}
\alias{makeMAR}
\title{Generate MAR Missingness Matrix}
\usage{
makeMAR(data, pm = NULL, cov = NULL, ignoreCols = NULL, threshold = NULL)
}
\arguments{
\item{data}{Data frame containing the dataset.}

\item{pm}{Proportion of MAR missingness.}

\item{cov}{Column index or name of the covariate.}

\item{ignoreCols}{Columns excluded from missingness.}

\item{threshold}{Threshold value defining eligibility for missingness.}
}
\value{
Logical matrix indicating values to be set to missing.
}
\description{
Creates a logical matrix indicating values that should be deleted
according to a missing-at-random (MAR) mechanism based on a covariate.
}
\keyword{internal}
