% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imposeMissing.R
\name{makeMCAR}
\alias{makeMCAR}
\title{Generate MCAR Missingness Matrix}
\usage{
makeMCAR(dims, pm = 0, cov = 0, ignoreCols = 0)
}
\arguments{
\item{dims}{Dimensions of the data matrix.}

\item{pm}{Proportion of missing completely at random.}

\item{cov}{Columns treated as covariates.}

\item{ignoreCols}{Columns excluded from missingness.}
}
\value{
Logical matrix indicating values to be set to missing.
}
\description{
Creates a logical matrix indicating values that should be deleted
according to a missing completely at random (MCAR) mechanism.
}
\keyword{internal}
