% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss.R
\name{miss}
\alias{miss}
\title{Specify Missing-Data Mechanisms for Simulation}
\usage{
miss(
  cov = 0,
  pmMCAR = 0,
  pmMAR = 0,
  logit = "",
  nforms = 0,
  itemGroups = list(),
  timePoints = 1,
  twoMethod = 0,
  prAttr = 0,
  m = 0,
  convergentCutoff = 0.8,
  ignoreCols = 0,
  threshold = 0,
  covAsAux = TRUE,
  logical = NULL,
  ...
)
}
\arguments{
\item{cov}{Numeric value indicating the proportion of covariates with missing data.}

\item{pmMCAR}{Proportion of data missing completely at random (MCAR).}

\item{pmMAR}{Proportion of data missing at random (MAR).}

\item{logit}{Character string specifying a logistic model used to generate
missingness probabilities.}

\item{nforms}{Number of planned missing forms.}

\item{itemGroups}{List defining groups of items used in planned missing designs.}

\item{timePoints}{Number of measurement occasions in longitudinal designs.}

\item{twoMethod}{Proportion of variables using a two-method measurement design.}

\item{prAttr}{Proportion of attributes used for generating missingness.}

\item{m}{Number of imputations (currently unused internally).}

\item{convergentCutoff}{Numeric cutoff used for assessing convergence when
imputation is performed.}

\item{ignoreCols}{Columns to ignore when imposing missingness.}

\item{threshold}{Threshold controlling missing-data generation.}

\item{covAsAux}{Logical indicating whether covariates should be treated as
auxiliary variables during estimation.}

\item{logical}{Optional logical matrix specifying a fixed missing-data pattern.}

\item{...}{Additional arguments passed internally to imputation procedures.}
}
\value{
An object of class \code{SimMissing} describing the missing-data
  mechanism used in simulations.
}
\description{
Creates a \code{SimMissing} object describing how missing data should be
imposed in simulated datasets. The object stores parameters controlling
missing-data mechanisms such as MCAR, MAR, planned missing designs, and
auxiliary-variable handling.
}
\details{
The resulting object is used internally by simulation functions to generate
missing data patterns before model estimation.


This function defines missing-data mechanisms used during simulation.
Currently, the default estimation method assumes full information maximum
likelihood (FIML) through \code{lavaan}.

Planned missing-data designs and logistic missingness mechanisms can be
specified using the corresponding arguments.
}
\seealso{
imposeMissing
}
