% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.lavaan}
\alias{model.lavaan}
\title{Convert a lavaan Model to a simsem Template}
\usage{
model.lavaan(
  object,
  std = FALSE,
  LY = NULL,
  PS = NULL,
  RPS = NULL,
  TE = NULL,
  RTE = NULL,
  BE = NULL,
  VTE = NULL,
  VY = NULL,
  VPS = NULL,
  VE = NULL,
  TY = NULL,
  AL = NULL,
  MY = NULL,
  ME = NULL,
  KA = NULL,
  GA = NULL
)
}
\arguments{
\item{object}{A fitted \code{lavaan} object.}

\item{std}{Logical indicating whether standardized parameters should be
used.}

\item{LY}{Factor loading matrix.}

\item{PS}{Factor covariance matrix.}

\item{RPS}{Factor correlation matrix.}

\item{TE}{Measurement error covariance matrix.}

\item{RTE}{Measurement error correlation matrix.}

\item{BE}{Structural regression coefficient matrix.}

\item{VTE}{Measurement error variances.}

\item{VY}{Total indicator variances.}

\item{VPS}{Factor variances.}

\item{VE}{Total latent variances.}

\item{TY}{Indicator intercepts.}

\item{AL}{Latent intercepts.}

\item{MY}{Indicator means.}

\item{ME}{Latent means.}

\item{KA}{Regression coefficients from covariates to indicators.}

\item{GA}{Regression coefficients from covariates to latent variables.}
}
\value{
A \code{SimSem} object.
}
\description{
Converts a fitted \code{lavaan} object into a \code{SimSem} template that
can be used for simulation with \code{simsem}.
}
