% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\alias{model.cfa}
\alias{model.path}
\alias{model.sem}
\title{Specify a Simulation Model Using lavaan Syntax Components}
\usage{
model(
  LY = NULL,
  PS = NULL,
  RPS = NULL,
  TE = NULL,
  RTE = NULL,
  BE = NULL,
  VTE = NULL,
  VY = NULL,
  VPS = NULL,
  VE = NULL,
  TY = NULL,
  AL = NULL,
  MY = NULL,
  ME = NULL,
  KA = NULL,
  GA = NULL,
  modelType = NULL,
  indLab = NULL,
  facLab = NULL,
  covLab = NULL,
  groupLab = "group",
  ngroups = 1,
  con = NULL
)

model.cfa(
  LY = NULL,
  PS = NULL,
  RPS = NULL,
  TE = NULL,
  RTE = NULL,
  VTE = NULL,
  VY = NULL,
  VPS = NULL,
  VE = NULL,
  TY = NULL,
  AL = NULL,
  MY = NULL,
  ME = NULL,
  KA = NULL,
  GA = NULL,
  indLab = NULL,
  facLab = NULL,
  covLab = NULL,
  groupLab = "group",
  ngroups = 1,
  con = NULL
)

model.path(
  PS = NULL,
  RPS = NULL,
  BE = NULL,
  VPS = NULL,
  VE = NULL,
  AL = NULL,
  ME = NULL,
  KA = NULL,
  GA = NULL,
  indLab = NULL,
  facLab = NULL,
  covLab = NULL,
  groupLab = "group",
  ngroups = 1,
  con = NULL
)

model.sem(
  LY = NULL,
  PS = NULL,
  RPS = NULL,
  TE = NULL,
  RTE = NULL,
  BE = NULL,
  VTE = NULL,
  VY = NULL,
  VPS = NULL,
  VE = NULL,
  TY = NULL,
  AL = NULL,
  MY = NULL,
  ME = NULL,
  KA = NULL,
  GA = NULL,
  indLab = NULL,
  facLab = NULL,
  covLab = NULL,
  groupLab = "group",
  ngroups = 1,
  con = NULL
)
}
\arguments{
\item{LY}{Factor loading matrix.}

\item{PS}{Factor covariance matrix.}

\item{RPS}{Factor correlation matrix.}

\item{TE}{Measurement error covariance matrix.}

\item{RTE}{Measurement error correlation matrix.}

\item{BE}{Structural regression coefficient matrix.}

\item{VTE}{Measurement error variances.}

\item{VY}{Total indicator variances.}

\item{VPS}{Factor variances.}

\item{VE}{Total latent variances.}

\item{TY}{Indicator intercepts.}

\item{AL}{Latent intercepts.}

\item{MY}{Indicator means.}

\item{ME}{Latent means.}

\item{KA}{Regression coefficients from covariates to indicators.}

\item{GA}{Regression coefficients from covariates to latent variables.}

\item{modelType}{Character string specifying the model type:
\code{"cfa"}, \code{"sem"}, or \code{"path"}.}

\item{indLab}{Optional indicator labels.}

\item{facLab}{Optional latent variable labels.}

\item{covLab}{Optional covariate labels.}

\item{groupLab}{Name of the grouping variable.}

\item{ngroups}{Number of groups in the model.}

\item{con}{Optional parameter constraints.}
}
\value{
A \code{SimSem} object containing the model specification and
simulation templates.
}
\description{
Creates a \code{SimSem} object representing a structural equation model
for simulation using \code{lavaan}. The model can be specified using
\code{SimMatrix} or \code{SimVector} objects describing factor loadings,
covariances, regressions, and other SEM components.
}
\details{
The function supports CFA, SEM, and path models, as well as multi-group
models. The returned object contains the parameter table used for estimation and
templates for generating simulated data.


Models are specified using \code{SimMatrix} and \code{SimVector} objects
that define the structure of loadings, covariances, regressions, and means.

The function automatically constructs the parameter table required by
\code{lavaan} and generates templates used for simulating data.

Multi-group models can be specified by supplying lists of matrices or by
setting \code{ngroups > 1}.
}
\seealso{
\code{\link{model.cfa}}, \code{\link{model.sem}},
\code{\link{model.path}}, \code{\link{estmodel}}
}
