% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPowerFitNested.R
\name{multipleAllEqual}
\alias{multipleAllEqual}
\title{Test Whether All Objects Are Equal}
\usage{
multipleAllEqual(...)
}
\arguments{
\item{...}{Objects to be compared.}
}
\value{
Logical value. Returns \code{TRUE} if all objects are equal and
  \code{FALSE} otherwise.
}
\description{
Test whether all supplied objects are equal. The comparison is based on
\code{\link{all.equal}}.
}
\examples{
multipleAllEqual(1:5, 1:5, seq(2, 10, 2) / 2) # Should be TRUE
multipleAllEqual(1:5, 1:6, seq(2, 10, 2) / 2) # Should be FALSE

}
