% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPowerFitNested.R
\name{multipleAllEqualList}
\alias{multipleAllEqualList}
\title{Check Equality of Objects in a List}
\usage{
multipleAllEqualList(obj)
}
\arguments{
\item{obj}{A list of objects.}
}
\value{
Logical value indicating whether all objects are equal.
}
\description{
Internal helper that checks whether all elements of a list are equal
using \code{all.equal()}.
}
\keyword{internal}
