% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPowerFitNested.R
\name{multipleAnyEqual}
\alias{multipleAnyEqual}
\title{Check Whether Any Objects Are Equal}
\usage{
multipleAnyEqual(...)
}
\arguments{
\item{...}{Objects to compare.}
}
\value{
Logical value indicating whether any objects are equal.
}
\description{
Determines whether any pair of supplied objects are equal using
\code{all.equal()}.
}
\keyword{internal}
