% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPowerFitNested.R
\name{multipleAnyEqualList}
\alias{multipleAnyEqualList}
\title{Check Whether Any Elements in a List Are Equal}
\usage{
multipleAnyEqualList(obj)
}
\arguments{
\item{obj}{A list of objects.}
}
\value{
Logical value indicating whether any objects are equal.
}
\description{
Internal helper that checks whether any pair of elements in a list
are equal using \code{all.equal()}.
}
\keyword{internal}
