\name{nullMatrix-class}
\Rdversion{1.1}
\docType{class}
\alias{nullMatrix-class}

\title{Class \code{"nullMatrix"}}
\description{
Null Matrix (Internal)
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("nullMatrix", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{.Data}:}{No element in it}
  }
}
\section{Methods}{
No methods defined with class "nullMatrix" in the signature.
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)
}
\note{
	Internal Class
}
