\name{nullSimMatrix-class}
\Rdversion{1.1}
\docType{class}
\alias{nullSimMatrix-class}
\alias{nullSymMatrix-class}
\alias{nullSimVector-class}
\alias{run,nullSimMatrix-method}
\alias{run,nullSymMatrix-method}
\alias{run,nullSimVector-method}

\title{Class Null object of \code{simMatrix}, \code{symMatrix}, and \code{simVector} classes (Internal)}
\description{
Represent null object of simMatrix class
}
\section{Objects from the Class}{
Cannot create from user interface.
}
\section{Slots}{
  \describe{
    \item{\code{Data}:}{Always NaN}
    \item{\code{Labels}:}{Always NaN}

  }
}
\section{Methods}{
  \describe{
    \item{run}{Return Null Matrix or Vector}
	 }
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)
}
\note{
Internal Class
}
\keyword{classes}
