\name{nullVector-class}
\Rdversion{1.1}
\docType{class}
\alias{nullVector-class}

\title{Class \code{"nullVector"}}
\description{
Null Vector (Internal)
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("nullVector", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{.Data}:}{No element in it}
  }
}
\section{Extends}{
Class \code{"\linkS4class{vector}"}, from data part.
}
\section{Methods}{
No methods defined with class "nullVector" in the signature.
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)
}
\note{
	Internal Class
}
