% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPowerFit.R
\name{overlapHist}
\alias{overlapHist}
\title{Plot overlapping histograms}
\usage{
overlapHist(
  a,
  b,
  colors = c("red", "blue", "purple"),
  breaks = NULL,
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  xlab = NULL,
  swap = FALSE
)
}
\arguments{
\item{a}{Data for the first histogram.}

\item{b}{Data for the second histogram.}

\item{colors}{Colors for the first histogram, second histogram,
and overlapping regions.}

\item{breaks}{Optional break specification.}

\item{xlim}{Range of the x-axis.}

\item{ylim}{Range of the y-axis.}

\item{main}{Plot title.}

\item{xlab}{X-axis label.}

\item{swap}{If \code{TRUE}, plot \code{b} first then \code{a}.}
}
\value{
None. Produces histograms.
}
\description{
Plot two overlapping histograms and highlight overlapping regions.
}
\details{
Original implementation provided by Chris Miller:
\url{http://chrisamiller.com/science/2010/07/20/transparent-overlapping-histograms-in-r/}.
Modified by Sunthud Pornprasertmanit.
}
\keyword{internal}
