% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{parsePopulation}
\alias{parsePopulation}
\title{Parse population parameter values}
\usage{
parsePopulation(paramSet, draws, group = 1, std = FALSE, covData = NULL)
}
\description{
Internal function used to compute population parameter values
based on generated parameter matrices.
}
\keyword{internal}
