% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imposeMissing.R
\name{parseSyntaxLogitMiss}
\alias{parseSyntaxLogitMiss}
\title{Parse Logistic Missingness Syntax}
\usage{
parseSyntaxLogitMiss(script)
}
\arguments{
\item{script}{Character string specifying logistic missingness syntax.}
}
\value{
Parsed model expressions.
}
\description{
Parses the syntax used to specify logistic missingness models.
}
\keyword{internal}
