% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imposeMissing.R
\name{permn}
\alias{permn}
\title{Generate All Permutations}
\usage{
permn(x, fun = NULL, ...)
}
\arguments{
\item{x}{Vector of elements.}

\item{fun}{Optional function applied to permutations.}
}
\value{
List of permutations.
}
\description{
Internal implementation of permutation generation used in
planned missing designs.
}
\keyword{internal}
