% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imposeMissing.R
\name{plannedMissing}
\alias{plannedMissing}
\title{Generate Planned Missingness Matrix}
\usage{
plannedMissing(
  dims = c(0, 0),
  nforms = NULL,
  itemGroups = NULL,
  twoMethod = NULL,
  cov = NULL,
  timePoints = 1,
  ignoreCols = NULL
)
}
\arguments{
\item{dims}{Dimensions of the dataset.}

\item{nforms}{Number of forms in the planned missing design.}

\item{itemGroups}{Optional list defining item groups.}

\item{twoMethod}{Two-method planned missing specification.}

\item{cov}{Covariate columns.}

\item{timePoints}{Number of measurement occasions.}

\item{ignoreCols}{Columns excluded from missingness.}
}
\value{
Logical matrix indicating values to be set to missing.
}
\description{
Creates a logical matrix representing planned missing data designs
such as matrix sampling or multi-form designs.
}
\keyword{internal}
