% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCutoff.R
\name{plot3DQtile}
\alias{plot3DQtile}
\title{Plot quantile surfaces for two predictors}
\usage{
plot3DQtile(
  x,
  y,
  z,
  df = 0,
  qtile = 0.5,
  useContour = TRUE,
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  main = NULL
)
}
\description{
Internal function that builds contour or perspective plots of predicted
quantiles from quantile regression models.
}
\keyword{internal}
