% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCutoff.R
\name{plotCutoff}
\alias{plotCutoff}
\title{Plot sampling distributions of fit indices with cutoff values}
\usage{
plotCutoff(
  object,
  alpha = NULL,
  revDirec = FALSE,
  usedFit = NULL,
  useContour = TRUE
)
}
\arguments{
\item{object}{A \code{\linkS4class{SimResult}} object containing simulation
results.}

\item{alpha}{Significance level used to derive cutoff values.}

\item{revDirec}{Logical indicating whether the direction of the cutoff
should be reversed. By default, the cutoff is placed on the side that
indicates worse fit (e.g., the right side of RMSEA or the left side of
CFI).}

\item{usedFit}{Character vector specifying which fit indices should
be plotted.}

\item{useContour}{Logical indicating whether contour plots should be
used when two varying parameters are present. If \code{FALSE},
perspective plots are produced instead.}
}
\value{
No value is returned. This function produces plots.
}
\description{
Plot sampling distributions of fit indices from a
\code{\linkS4class{SimResult}} object. Optional cutoff values can be
added either by specifying an \code{alpha} level or by supplying
predefined cutoffs.
}
\details{
If simulation conditions vary (e.g., sample size or missing data rates),
the function visualizes how cutoff values change across these conditions.
}
\examples{
\dontrun{
loading <- matrix(0, 6, 2)
loading[1:3, 1] <- NA
loading[4:6, 2] <- NA

loadingValues <- matrix(0, 6, 2)
loadingValues[1:3, 1] <- 0.7
loadingValues[4:6, 2] <- 0.7

LY <- bind(loading, loadingValues)

latent.cor <- matrix(NA, 2, 2)
diag(latent.cor) <- 1

RPS <- binds(latent.cor, 0.5)

error.cor <- matrix(0, 6, 6)
diag(error.cor) <- 1

RTE <- binds(error.cor)

CFA.Model <- model(LY = LY, RPS = RPS, RTE = RTE, modelType = "CFA")

Output <- sim(5, n = 200, model = CFA.Model)

plotCutoff(Output, 0.05, usedFit = c("RMSEA", "SRMR", "CFI", "TLI"))

Output2 <- sim(NULL, n = seq(450, 500, 10), model = CFA.Model)

plotCutoff(Output2, 0.05)

Output3 <- sim(NULL,
  n = seq(450, 500, 10),
  pmMCAR = c(0, 0.05, 0.1, 0.15),
  model = CFA.Model
)

plotCutoff(Output3, 0.05)
}

}
\seealso{
\itemize{
\item \code{\linkS4class{SimResult}} for simulation result objects.
\item \code{\link{getCutoff}} to compute cutoff values based on
null-hypothesis sampling distributions.
}
}
