% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCutoffNonNested.R
\name{plotCutoffNonNested}
\alias{plotCutoffNonNested}
\title{Plot cutoff distributions for non-nested model comparisons}
\usage{
plotCutoffNonNested(
  dat1Mod1,
  dat1Mod2,
  dat2Mod1 = NULL,
  dat2Mod2 = NULL,
  alpha = 0.05,
  cutoff = NULL,
  usedFit = NULL,
  useContour = T,
  onetailed = FALSE
)
}
\arguments{
\item{dat1Mod1}{A \code{\linkS4class{SimResult}} object containing results
from fitting Model 1 to datasets generated from Model 1.}

\item{dat1Mod2}{A \code{\linkS4class{SimResult}} object containing results
from fitting Model 2 to datasets generated from Model 1.}

\item{dat2Mod1}{A \code{\linkS4class{SimResult}} object containing results
from fitting Model 1 to datasets generated from Model 2.}

\item{dat2Mod2}{A \code{\linkS4class{SimResult}} object containing results
from fitting Model 2 to datasets generated from Model 2.}

\item{alpha}{Significance level used to compute cutoff values.}

\item{cutoff}{A numeric vector of predefined cutoff values for the
differences in fit indices.}

\item{usedFit}{Character vector specifying which fit indices should be
plotted.}

\item{useContour}{Logical indicating whether contour plots should be used
when two varying parameters are present. If \code{FALSE}, perspective plots
are produced instead.}

\item{onetailed}{Logical indicating whether the cutoff should be derived
from a one-tailed test. If \code{FALSE}, a two-tailed test is used.}
}
\value{
No value is returned. This function produces plots.
}
\description{
Plot sampling distributions of the differences in fit indices between
two non-nested models. Optional cutoff values can be added either by
specifying an \code{alpha} level or by supplying predefined cutoffs.
}
\examples{
\dontrun{
# Model A: Factor 1 on Items 1-3 and Factor 2 on Items 4-8
loading.A <- matrix(0, 8, 2)
loading.A[1:3, 1] <- NA
loading.A[4:8, 2] <- NA

LY.A <- bind(loading.A, 0.7)

latent.cor <- matrix(NA, 2, 2)
diag(latent.cor) <- 1

RPS <- binds(latent.cor, "runif(1, 0.7, 0.9)")
RTE <- binds(diag(8))

CFA.Model.A <- model(LY = LY.A, RPS = RPS, RTE = RTE, modelType = "CFA")

# Model B: Factor 1 on Items 1-4 and Factor 2 on Items 5-8
loading.B <- matrix(0, 8, 2)
loading.B[1:4, 1] <- NA
loading.B[5:8, 2] <- NA

LY.B <- bind(loading.B, 0.7)

CFA.Model.B <- model(LY = LY.B, RPS = RPS, RTE = RTE, modelType = "CFA")

Output.A.A <- sim(10, n = 500, model = CFA.Model.A, generate = CFA.Model.A)
Output.A.B <- sim(10, n = 500, model = CFA.Model.B, generate = CFA.Model.A)
Output.B.A <- sim(10, n = 500, model = CFA.Model.A, generate = CFA.Model.B)
Output.B.B <- sim(10, n = 500, model = CFA.Model.B, generate = CFA.Model.B)

# Plot cutoffs for both models
plotCutoffNonNested(Output.A.A, Output.A.B, Output.B.A, Output.B.B)

# Plot cutoffs for Model A only
plotCutoffNonNested(Output.A.A, Output.A.B)

# Plot cutoffs using a one-tailed test
plotCutoffNonNested(Output.A.A, Output.A.B, onetailed = TRUE)
}

}
\seealso{
\itemize{
\item \code{\linkS4class{SimResult}} for simulation result objects used
in this function.
\item \code{\link{getCutoffNonNested}} to compute cutoff values for
non-nested model comparisons.
}
}
