% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDist.R
\name{plotDist}
\alias{plotDist}
\title{Plot a distribution from a SimDataDist object}
\usage{
plotDist(object, xlim = NULL, ylim = NULL, r = 0, var = NULL, contour = TRUE)
}
\arguments{
\item{object}{A \code{\linkS4class{SimDataDist}} object describing
the data distribution to plot.}

\item{xlim}{Numeric vector of length two specifying the limits of the
x-axis.}

\item{ylim}{Numeric vector of length two specifying the limits of the
y-axis. Only used for joint distributions of two variables.}

\item{r}{Correlation between the two variables when plotting a joint
distribution.}

\item{var}{Vector of variable indices to plot. The length of the vector
cannot exceed two.}

\item{contour}{Logical indicating whether a contour plot should be used
when plotting two variables. If \code{FALSE}, a perspective plot is used.}
}
\value{
No value is returned. This function produces plots.
}
\description{
Plot marginal or joint distributions from a
\code{\linkS4class{SimDataDist}} object.
}
\details{
If one variable is selected, a marginal density is plotted.
If two variables are selected, a joint distribution is plotted
using either a contour plot or a perspective plot.
}
\examples{
datadist <- bindDist(skewness = c(0, -2, 2), kurtosis = c(2, 4, 4))

# Plot the joint distribution of variables 1 and 2
plotDist(datadist, r = 0.5, var = 1:2)

# Plot the marginal distribution of variable 3
plotDist(datadist, var = 3)

\dontrun{
datadist2 <- bindDist(
  c("chisq", "t", "f"),
  list(df = 5),
  list(df = 3),
  list(df1 = 3, df2 = 5)
)

plotDist(datadist2, r = 0.5, var = 1:2)
plotDist(datadist2, var = 3)
}

}
\seealso{
\itemize{
\item \code{\linkS4class{SimDataDist}} for creating distribution objects.
}
}
