% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDist.R
\name{plotDist1D}
\alias{plotDist1D}
\title{Plot a marginal distribution}
\usage{
plotDist1D(distName, param, xlim = NULL, reverse = FALSE)
}
\description{
Internal function used by \code{\link{plotDist}} to plot
one-dimensional marginal distributions.
}
\keyword{internal}
