% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDist.R
\name{plotDist2D}
\alias{plotDist2D}
\title{Plot a two-dimensional joint distribution}
\usage{
plotDist2D(object, xlim = NULL, ylim = NULL, r = 0, contour = TRUE)
}
\description{
Internal function used by \code{\link{plotDist}} to plot
joint distributions of two variables using contour or
perspective plots.
}
\keyword{internal}
