% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPowerFit.R
\name{plotLogisticFit}
\alias{plotLogisticFit}
\title{Plot logistic curves predicting rejection of misspecified models}
\usage{
plotLogisticFit(
  altObject,
  nullObject = NULL,
  cutoff = NULL,
  usedFit = NULL,
  x,
  xval,
  alpha = 0.05,
  useContour = TRUE,
  df = 0
)
}
\arguments{
\item{altObject}{A \code{data.frame} containing fit-index distributions when
the hypothesized model is \code{FALSE}.}

\item{nullObject}{A \code{data.frame} containing fit-index distributions when
the hypothesized model is \code{TRUE}.}

\item{cutoff}{A vector of a priori fit-index cutoffs.}

\item{usedFit}{Fit indices to plot.}

\item{x}{Predictor values.}

\item{xval}{Predictor grid values used for prediction.}

\item{alpha}{Significance level.}

\item{useContour}{Whether contour plots should be used for 2D predictors.}

\item{df}{Degrees of freedom for spline predictors.}
}
\value{
None. Produces plots.
}
\description{
This function determines whether a hypothesized model is rejected in each
dataset and plots logistic curves predicting rejection probabilities from
predictor values.
}
\keyword{internal}
