% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imposeMissing.R
\name{plotLogitMiss}
\alias{plotLogitMiss}
\title{Visualize Missing Proportion from Logistic Missingness Models}
\usage{
plotLogitMiss(
  script,
  ylim = c(0, 1),
  x1lim = c(-3, 3),
  x2lim = c(-3, 3),
  otherx = 0,
  useContour = TRUE
)
}
\arguments{
\item{script}{The script used in specifying missing data using the
logistic regression model. See further details in the \code{logit}
argument of the \code{\link{miss}} function.}

\item{ylim}{The range of missing proportion to be plotted.}

\item{x1lim}{The range of the first independent variable to be plotted.}

\item{x2lim}{The range of the second independent variable to be plotted.}

\item{otherx}{The value used to fix additional independent variables.}

\item{useContour}{If two independent variables are present, the function
will produce a 3D representation. By default a contour plot is used.
If \code{FALSE}, a perspective plot is produced instead.}
}
\value{
This function does not return a value. It produces a plot illustrating
the missingness probability implied by the logistic model.

* If the number of independent variables is 0, a bar plot is produced.
* If the number of independent variables is 1, a logistic curve is plotted.
* If the number of independent variables is 2, a contour or perspective plot
  is produced.
}
\description{
Visualize the missing proportion when the logistic regression method
is used. The maximum number of independent variables supported for
visualization is two. Any additional independent variables will be
fixed at a specified value (default = 0).
}
\examples{
script <- "y1 ~ 0.05 + 0.1*y2 + 0.3*y3
y4 ~ -2 + 0.1*y4
y5 ~ -0.5"
plotLogitMiss(script)

script2 <- "y1 ~ 0.05 + 0.5*y3
y2 ~ p(0.2)
y3 ~ p(0.1) + -1*y1
y4 ~ p(0.3) + 0.2*y1 + -0.3*y2
y5 ~ -0.5"
plotLogitMiss(script2)

}
\seealso{
\code{\link{miss}} to create the missing data template \cr
\code{\link{impose}} to impose missing data
}
