% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMisfit.R
\name{plotMisfit}
\alias{plotMisfit}
\title{Plot population misfit in a simulation result object}
\usage{
plotMisfit(object, usedFit = "default", misParam = NULL)
}
\arguments{
\item{object}{A \code{\linkS4class{SimResult}} object.}

\item{usedFit}{Sample fit indices or population misfit indices used in the
plot. All sample fit indices are available. The available population misfit
indices are `"pop.f0"`, `"pop.rmsea"`, and `"pop.srmr"`.

If \code{misParam} is not specified, all population misfit indices are used.
If \code{misParam} is specified, `"pop.rmsea"` is used by default.}

\item{misParam}{Index or name of misspecified parameters to use in the plot.}
}
\value{
No value is returned. This function produces plots.
}
\description{
Plot either the distribution of population misfit or the relationship
between misspecified parameters and population misfit (or sample fit indices)
from a \code{\linkS4class{SimResult}} object.
}
\details{
If no misspecification parameter is specified, the function plots histograms
of population misfit indices. If misspecification parameters are specified,
the function plots scatterplots showing the relationship between the
misspecification values and the selected fit index.
}
\examples{
path.BE <- matrix(0, 4, 4)
path.BE[3, 1:2] <- NA
path.BE[4, 3] <- NA

starting.BE <- matrix("", 4, 4)
starting.BE[3, 1:2] <- "runif(1, 0.3, 0.5)"
starting.BE[4, 3] <- "runif(1, 0.5, 0.7)"

mis.path.BE <- matrix(0, 4, 4)
mis.path.BE[4, 1:2] <- "runif(1, -0.1, 0.1)"

BE <- bind(path.BE, starting.BE, misspec = mis.path.BE)

residual.error <- diag(4)
residual.error[1,2] <- residual.error[2,1] <- NA
RPS <- binds(residual.error, "rnorm(1, 0.3, 0.1)")

ME <- bind(rep(NA, 4), 0)

Path.Model <- model(RPS = RPS, BE = BE, ME = ME, modelType = "Path")

# In real applications, more replications are recommended
Output <- sim(20, n = 500, Path.Model)

# Plot distribution of population misfit
plotMisfit(Output)

# Plot relationship between population RMSEA and misspecified paths
plotMisfit(Output, misParam = 1:2)

# Plot relationship between sample CFI and misspecified paths
plotMisfit(Output, usedFit = "CFI", misParam = 1:2)

}
\seealso{
\code{\linkS4class{SimResult}}
}
