% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPowerFit.R
\name{plotPowerFit}
\alias{plotPowerFit}
\title{Plot sampling distributions of fit indices that visualize power of rejecting datasets underlying misspecified models}
\usage{
plotPowerFit(
  altObject,
  nullObject = NULL,
  cutoff = NULL,
  usedFit = NULL,
  alpha = 0.05,
  contN = TRUE,
  contMCAR = TRUE,
  contMAR = TRUE,
  useContour = TRUE,
  logistic = TRUE
)
}
\arguments{
\item{altObject}{A \code{\linkS4class{SimResult}} object saving the simulation
results of fitting the hypothesized model when the hypothesized model is
\code{FALSE}.}

\item{nullObject}{A \code{\linkS4class{SimResult}} object saving the simulation
results of fitting the hypothesized model when the hypothesized model is
\code{TRUE}. This argument may not be specified if the \code{cutoff} argument
is provided.}

\item{cutoff}{A vector of a priori cutoffs for fit indices.}

\item{usedFit}{Vector of names of fit indices that researchers wish to plot.}

\item{alpha}{A priori alpha level.}

\item{contN}{Include varying sample size in the power plot if available.}

\item{contMCAR}{Include varying MCAR (missing completely at random percentage)
in the power plot if available.}

\item{contMAR}{Include varying MAR (missing at random percentage) in the
power plot if available.}

\item{useContour}{If two of sample size, percent completely at random, and
percent missing at random vary, a 3D plot will be produced. A contour graph
is the default. If \code{FALSE}, a perspective plot is used.}

\item{logistic}{If \code{TRUE} and varying parameters exist (e.g., sample size
or percent missing), a logistic regression curve predicting significance from
the varying parameters is used. If \code{FALSE}, an overlaying scatterplot
with a cutoff line is plotted.}
}
\value{
None. This function only produces plots.
}
\description{
This function plots sampling distributions of fit indices that visualize
power in rejecting misspecified models.
}
\examples{
\dontrun{
# Null model: One-factor model
loading.null <- matrix(0, 6, 1)
loading.null[1:6, 1] <- NA
LY.NULL <- bind(loading.null, 0.7)
RPS.NULL <- binds(diag(1))
RTE <- binds(diag(6))
CFA.Model.NULL <- model(LY = LY.NULL, RPS = RPS.NULL, RTE = RTE, modelType="CFA")

Output.NULL <- sim(50, n=50, model=CFA.Model.NULL, generate=CFA.Model.NULL)

# Alternative model: Two-factor model
loading.alt <- matrix(0, 6, 2)
loading.alt[1:3, 1] <- NA
loading.alt[4:6, 2] <- NA
LY.ALT <- bind(loading.alt, 0.7)
latent.cor.alt <- matrix(NA, 2, 2)
diag(latent.cor.alt) <- 1
RPS.ALT <- binds(latent.cor.alt, 0.5)
CFA.Model.ALT <- model(LY = LY.ALT, RPS = RPS.ALT, RTE = RTE, modelType="CFA")

Output.ALT <- sim(50, n=50, model=CFA.Model.NULL, generate=CFA.Model.ALT)

plotPowerFit(Output.ALT, nullObject=Output.NULL,
             usedFit=c("RMSEA","CFI","TLI","SRMR"))
}

}
\seealso{
\itemize{
\item \code{\linkS4class{SimResult}} for simulation results used in this function.
\item \code{\link{getCutoff}} to compute cutoff values based on null hypothesis
sampling distributions.
\item \code{\link{getPowerFit}} to compute power for rejecting a hypothesized model.
}
}
