% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPowerFit.R
\name{plotPowerFitDf}
\alias{plotPowerFitDf}
\title{Plot sampling distributions of fit indices that visualize power of rejecting misspecified models (data.frame version)}
\usage{
plotPowerFitDf(
  altObject,
  nullObject = NULL,
  cutoff = NULL,
  usedFit = NULL,
  alpha = 0.05,
  x = NULL,
  xval = NULL,
  useContour = TRUE,
  logistic = TRUE
)
}
\arguments{
\item{altObject}{A \code{data.frame} containing fit-index distributions when
the hypothesized model is \code{FALSE}.}

\item{nullObject}{A \code{data.frame} containing fit-index distributions when
the hypothesized model is \code{TRUE}. This argument may be omitted if
\code{cutoff} is specified.}

\item{cutoff}{A vector of a priori cutoffs for fit indices.}

\item{usedFit}{Vector of names of fit indices to plot.}

\item{alpha}{A priori alpha level.}

\item{x}{Predictor values used to evaluate power.}

\item{xval}{Predictor values for plotting predicted cutoffs.}

\item{useContour}{If two predictors vary, a contour plot is used by default.
If \code{FALSE}, a perspective plot is used.}

\item{logistic}{Logical indicating whether logistic regression smoothing
should be used.}
}
\value{
None. This function only produces plots.
}
\description{
This function plots sampling distributions of fit indices that visualize
power in rejecting misspecified models. This function is similar to
\code{\link{plotPowerFit}} but uses \code{data.frame} inputs instead of
\code{\linkS4class{SimResult}} objects.
}
\keyword{internal}
