% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPowerCoverage.R
\name{plotPowerSig}
\alias{plotPowerSig}
\title{Plot power curves from significance indicators}
\usage{
plotPowerSig(sig, x = NULL, xval = NULL, mainName = NULL, useContour = TRUE)
}
\description{
Internal function that plots predicted probabilities of significance
based on logistic regression given a matrix of significance indicators
and predictor variables.
}
\keyword{internal}
