% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCutoff.R
\name{plotQtile}
\alias{plotQtile}
\title{Plot conditional quantile curves}
\usage{
plotQtile(x, y, df = 0, qtile = NULL, ...)
}
\description{
Internal helper function used to visualize conditional quantiles
estimated via quantile regression.
}
\keyword{internal}
