% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPowerFit.R
\name{plotScatter}
\alias{plotScatter}
\title{Plot overlaying scatter plots visualizing rejection power}
\usage{
plotScatter(
  altObject,
  nullObject = NULL,
  cutoff = NULL,
  usedFit = NULL,
  x,
  alpha = 0.05,
  df = 5
)
}
\arguments{
\item{altObject}{A \code{data.frame} containing fit-index distributions when
the hypothesized model is \code{FALSE}.}

\item{nullObject}{A \code{data.frame} containing fit-index distributions when
the hypothesized model is \code{TRUE}.}

\item{cutoff}{A vector of a priori cutoffs.}

\item{usedFit}{Fit indices to plot.}

\item{x}{Predictor values.}

\item{alpha}{Significance level.}

\item{df}{Degrees of freedom for spline predictors.}
}
\value{
None. Produces scatter plots.
}
\description{
Plot fit-index values against predictor values. The plot may include
alternative distributions, null distributions, and cutoff curves.
}
\keyword{internal}
