% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawParam.R
\name{popDiscrepancy}
\alias{popDiscrepancy}
\title{Population discrepancy function}
\usage{
popDiscrepancy(paramM, paramCM, misspecM, misspecCM)
}
\arguments{
\item{paramM}{Population mean vector under the correctly specified model.}

\item{paramCM}{Population covariance matrix under the correctly specified
model.}

\item{misspecM}{Population mean vector under the misspecified model.}

\item{misspecCM}{Population covariance matrix under the misspecified model.}
}
\value{
Numeric value of the population discrepancy function.
}
\description{
Computes the population discrepancy function between two mean and
covariance structures.
}
\details{
This corresponds to the population discrepancy due to approximation
described by Browne and Cudeck (1992).
}
\keyword{internal}
