% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{popMisfitParams}
\alias{popMisfitParams}
\title{Compute population misfit measures}
\usage{
popMisfitParams(psl, df = NULL, covData = NULL)
}
\description{
Computes population misfit measures based on generated model
parameters.
}
\keyword{internal}
