% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValue.R
\name{pValueCondCutoff}
\alias{pValueCondCutoff}
\title{Find a p value when the target is conditional on predictor values}
\usage{
pValueCondCutoff(target, dist, revDirec = FALSE, x = NULL, xval = NULL, df = 0)
}
\arguments{
\item{target}{A target value used to find \code{p} values.}

\item{dist}{The comparison distribution.}

\item{revDirec}{Logical indicating whether to reverse the direction
of comparison.}

\item{x}{Data frame of predictor values. The number of rows must match
the number of rows in \code{dist}.}

\item{xval}{Predictor value at which the conditional p-value is evaluated.}

\item{df}{Degrees of freedom used in spline modeling. If \code{df = 0},
splines are not used.}
}
\value{
A vector of \emph{p}-values based on the comparison.
}
\description{
Find a \emph{p} value when the target value is conditional (valid)
only for a specific value of a predictor.
}
\details{
That is, the target value is applicable only at a given value of a
predictor, and the p-value is computed relative to that conditional
distribution.
}
\keyword{internal}
