% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValueNested.R
\name{pValueNested}
\alias{pValueNested}
\title{Find p-values (1 - percentile) for a nested model comparison}
\usage{
pValueNested(
  outNested,
  outParent,
  simNested,
  simParent,
  usedFit = NULL,
  nVal = NULL,
  pmMCARval = NULL,
  pmMARval = NULL,
  df = 0
)
}
\arguments{
\item{outNested}{A \code{lavaan}-class object saving the analysis result
of the nested model from the target dataset.}

\item{outParent}{A \code{lavaan}-class object saving the analysis result
of the parent model from the target dataset.}

\item{simNested}{A \code{\linkS4class{SimResult}} object containing the
analysis results of the nested model from multiple replications.}

\item{simParent}{A \code{\linkS4class{SimResult}} object containing the
analysis results of the parent model from multiple replications.}

\item{usedFit}{Vector of names of fit indices that researchers wish to
obtain cutoffs from. The default is to use all available fit indices.}

\item{nVal}{The sample size value for which researchers wish to find
the \emph{p}-value.}

\item{pmMCARval}{The percent missing completely at random value for which
researchers wish to find the \emph{p}-value.}

\item{pmMARval}{The percent missing at random value for which researchers
wish to find the \emph{p}-value.}

\item{df}{The degree of freedom used in spline methods for predicting fit
indices by predictors. If \code{df = 0}, the spline method is not applied.}
}
\value{
A vector of \emph{p}-values based on the comparison of the difference in
fit indices from the observed data with the simulation results.

The \emph{p}-values of the requested fit indices are returned along with
two additional values:

\itemize{
\item \code{andRule}: The proportion of replications in which all fit
indices indicate a better model than the observed data. This represents
the most stringent rule for retaining a hypothesized model.

\item \code{orRule}: The proportion of replications in which at least one
fit index indicates a better model than the observed data. This represents
the most lenient rule for retaining a hypothesized model.
}
}
\description{
This function provides \emph{p}-values by comparing the differences in
fit indices between nested models with the simulation results of both
the parent and nested models when the nested model is true.
}
\details{
In comparing fit indices, the \emph{p}-value is the proportion of
replications that provide less preference for the nested model
(e.g., larger negative difference in CFI values or larger positive
difference in RMSEA values) than the analysis result from the
observed data.


The \emph{p}-value is computed as the proportion of replications that
provide less preference for the nested model compared with the observed
data. For example, this may correspond to larger negative differences
in CFI values or larger positive differences in RMSEA values.
}
\examples{
\dontrun{
library(lavaan)

# Nested Model: Linear growth curve model
LY <- matrix(1, 4, 2)
LY[, 2] <- 0:3
PS <- matrix(NA, 2, 2)
TY <- rep(0, 4)
AL <- rep(NA, 2)
TE <- diag(NA, 4)
nested <- estmodel(
  LY = LY, PS = PS, TY = TY, AL = AL, TE = TE,
  modelType = "CFA", indLab = paste("t", 1:4, sep = "")
)

# Parent Model: Unconditional growth curve model
LY2 <- matrix(1, 4, 2)
LY2[, 2] <- c(0, NA, NA, 3)
parent <- estmodel(
  LY = LY2, PS = PS, TY = TY, AL = AL, TE = TE,
  modelType = "CFA", indLab = paste("t", 1:4, sep = "")
)

# Analyze the output
outNested <- analyze(nested, Demo.growth)
outParent <- analyze(parent, Demo.growth)

# Create data template from the nested model with small misfit
loadingMis <- matrix(0, 4, 2)
loadingMis[2:3, 2] <- "runif(1, -0.1, 0.1)"
datamodel <- model.lavaan(outNested, LY = loadingMis)

# Get the sample size
n <- nrow(Demo.growth)

# The actual replications should be much greater than 30
simNestedNested <- sim(30, n = n, nested, generate = datamodel)
simNestedParent <- sim(30, n = n, parent, generate = datamodel)

# Compare observed fit indices against simulated distributions
pValueNested(outNested, outParent, simNestedNested, simNestedParent)
}

}
\seealso{
\code{\linkS4class{SimResult}}
}
