% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValue.R
\name{pValueVariedCutoff}
\alias{pValueVariedCutoff}
\title{Compute p values for varying cutoff values}
\usage{
pValueVariedCutoff(
  cutoff,
  obtainedValue,
  revDirec = FALSE,
  x = NULL,
  xval = NULL
)
}
\arguments{
\item{cutoff}{Vector of cutoff values.}

\item{obtainedValue}{Comparison distribution.}

\item{revDirec}{Logical indicating whether the comparison direction
should be reversed.}

\item{x}{Predictor matrix.}

\item{xval}{Predictor value at which the p-value is evaluated.}
}
\value{
A vector of \emph{p}-values.
}
\description{
Compute a \emph{p} value when cutoff values vary across simulation
conditions, typically as a function of predictor variables.
}
\keyword{internal}
