% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{reduceMisspecSet}
\alias{reduceMisspecSet}
\title{Reduce misspecification parameter sets}
\usage{
reduceMisspecSet(
  misspecSet,
  latent,
  indLab = NULL,
  facLab = NULL,
  covLab = NULL
)
}
\description{
Internal function used to convert misspecification matrices into
named parameter vectors used in simulation summaries.
}
\keyword{internal}
