% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{renameExtraParam}
\alias{renameExtraParam}
\title{Rename extra parameters from constraint expressions}
\usage{
renameExtraParam(lhs, op, rhs, refpt = NULL)
}
\description{
Converts lavaan constraint syntax into readable parameter labels.
}
\keyword{internal}
