% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValue.R
\name{revText}
\alias{revText}
\title{Reverse the direction of a proportion value}
\usage{
revText(val)
}
\arguments{
\item{val}{Value or character representation of a proportion.}
}
\value{
The reversed value or text.
}
\description{
Reverse the proportion value by subtracting it from 1.
This function can also reverse text representations such as
\code{"> .98"} to \code{"< .02"}.
}
\keyword{internal}
