% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{runRep}
\alias{runRep}
\title{Run a single simulation replication}
\usage{
runRep(
  simConds,
  model,
  generateO = NULL,
  miss = NULL,
  datafun = NULL,
  lavaanfun = NULL,
  outfun = NULL,
  outfundata = NULL,
  facDist = NULL,
  indDist = NULL,
  indLab = NULL,
  errorDist = NULL,
  sequential = FALSE,
  saveLatentVar = FALSE,
  realData = NULL,
  covData = NULL,
  silent = FALSE,
  modelBoot = FALSE,
  maxDraw = 50,
  misfitType = "f0",
  misfitBounds = NULL,
  averageNumMisspec = NULL,
  optMisfit = NULL,
  optDraws = 50,
  createOrder = c(1, 2, 3),
  aux = NULL,
  paramOnly = FALSE,
  dataOnly = FALSE,
  smartStart = TRUE,
  popData = NULL,
  group = NULL,
  mxFit = FALSE,
  mxMixture = FALSE,
  citype = NULL,
  cilevel = 0.95,
  stopOnError = FALSE,
  ...
)
}
\value{
A list containing results from a single replication.
}
\description{
Internal function used by \code{\link{sim}} to execute a single
Monte Carlo replication. The function generates data, applies missingness,
analyzes the dataset, and extracts relevant results.
}
\keyword{internal}
