% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySimResult.R
\name{setPopulation}
\alias{setPopulation}
\title{Set population parameter values}
\usage{
setPopulation(target, population)
}
\arguments{
\item{target}{A \code{SimResult} object.}

\item{population}{A population model specification. This can be a
\code{SimSem} object, a lavaan object, a lavaan model syntax, or a parameter table.}
}
\value{
The updated \code{SimResult} object with population parameter values set.
}
\description{
This function extracts and sets population parameter values from a
population model specification.
}
