% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SimDataDist-class}
\alias{SimDataDist-class}
\alias{summary,SimDataDist-method}
\alias{plotDist,SimDataDist-method}
\title{Class "SimDataDist": Data distribution object}
\description{
This class will provide the distribution of a dataset.
}
\section{Slots}{

\describe{
\item{\code{p}}{Number of variables.}

\item{\code{margins}}{A character vector specifying all the marginal distributions.}

\item{\code{paramMargins}}{A list whose each component is a list of named components,
giving the parameter values of the marginal distributions.}

\item{\code{keepScale}}{Logical value indicating whether to transform the data to
maintain the scale.}

\item{\code{reverse}}{Vector indicating whether variables should be reversed.}

\item{\code{copula}}{Copula object specifying the dependence structure.}

\item{\code{skewness}}{Vector of skewness values.}

\item{\code{kurtosis}}{Vector of kurtosis values.}
}}

\section{Objects from the Class}{

Objects can be created by \code{\link{bindDist}} function. It can also be
called from the form \code{new("SimDataDist", ...)}.
}

