\name{simDist-class}
\Rdversion{1.1}
\docType{class}
\alias{simDist-class}
\title{Class "simDist"}
\description{
All distribution objects. (Virtual Class)
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "simDist" in the signature.
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; \email{psunthud@ku.edu})
}
\seealso{
	List of all distribution objects.
	\enumerate{
		\item \code{\linkS4class{Rnorm}}
		\item \code{\linkS4class{Runif}}
	}
}
\examples{
showClass("simDist")
}
\keyword{classes}
