% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SimMatrix-class}
\alias{SimMatrix-class}
\alias{summaryShort,SimMatrix-method}
\alias{summary,SimMatrix-method}
\title{Matrix object: Random parameters matrix}
\description{
This object can be used to represent a matrix in an SEM model. It contains free
parameters, fixed values, starting values, and model misspecification.
This object can represent mean, intercept, or variance matrices.
}
\section{Slots}{

\describe{
\item{\code{free}}{The free-parameter vector. Any NA elements or character elements
are free. Any numeric elements are fixed as the specified number. If any free
elements have the same characters (except NA), the elements are equally constrained.}

\item{\code{popParam}}{Population parameter values.}

\item{\code{misspec}}{Model misspecification values.}

\item{\code{symmetric}}{Logical value indicating whether the matrix is symmetric.}
}}

\section{Objects from the Class}{

This object is created by \code{\link{bind}} or \code{\link{binds}} function.
}

