% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SimMissing-class}
\alias{SimMissing-class}
\alias{summary,SimMissing-method}
\title{Class "SimMissing"}
\description{
Missing information imposing on the complete dataset.
}
\section{Slots}{

\describe{
\item{\code{cov}}{Column indices of any normally distributed covariates used in the data set.}

\item{\code{logit}}{The script used for imposing missing values by logistic regression.
See \code{\link{miss}} for further details.}

\item{\code{pmMCAR}}{Proportion of missingness to introduce completely at random across all variables.}

\item{\code{pmMAR}}{Proportion of missingness to introduce using the listed covariates as predictors.}

\item{\code{nforms}}{Number of forms for planned missing designs.}

\item{\code{itemGroups}}{List specifying item groups.}

\item{\code{twoMethod}}{Vector specifying two-method missingness.}

\item{\code{prAttr}}{Vector of probabilities of attrition.}

\item{\code{m}}{Number of imputations.}

\item{\code{package}}{Package used for missing data handling.}

\item{\code{args}}{Additional arguments passed to missing data functions.}

\item{\code{convergentCutoff}}{Convergence cutoff value.}

\item{\code{timePoints}}{Time points for longitudinal missingness.}

\item{\code{ignoreCols}}{Columns to ignore.}

\item{\code{threshold}}{Threshold values.}

\item{\code{covAsAux}}{Logical indicating whether covariates are treated as auxiliary variables.}

\item{\code{logical}}{Logical matrix specifying missing data patterns.}
}}

\section{Objects from the Class}{

Objects can be created by \code{\link{miss}} function.
}

