% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SimResult-class}
\alias{SimResult-class}
\alias{summary,SimResult-method}
\alias{summaryShort,SimResult-method}
\title{Class "SimResult": Simulation Result Object}
\description{
This class will save data analysis results from multiple replications, such
as fit indices cutoffs or power, parameter values, model misspecification,
etc.
}
\section{Slots}{

\describe{
\item{\code{modelType}}{Analysis model type (CFA, Path, or SEM).}

\item{\code{nRep}}{Number of replications used to generate and analyze simulated data.}

\item{\code{coef}}{Parameter estimates from each replication.}

\item{\code{se}}{Standard errors of parameter estimates from each replication.}

\item{\code{fit}}{Fit indices from each replication.}

\item{\code{converged}}{Logical vector indicating convergence status.}

\item{\code{paramValue}}{Population parameter values.}

\item{\code{stdParamValue}}{Standardized population parameter values.}

\item{\code{misspecValue}}{Model misspecification values.}

\item{\code{popFit}}{Population fit indices.}

\item{\code{FMI1}}{Fraction of missing information measure 1.}

\item{\code{FMI2}}{Fraction of missing information measure 2.}

\item{\code{cilower}}{Lower confidence interval bounds.}

\item{\code{ciupper}}{Upper confidence interval bounds.}

\item{\code{stdCoef}}{Standardized coefficients.}

\item{\code{stdSe}}{Standardized standard errors.}

\item{\code{seed}}{Random seed used in simulation.}

\item{\code{n}}{Sample size.}

\item{\code{nobs}}{Number of observed cases.}

\item{\code{pmMCAR}}{Percent missing completely at random.}

\item{\code{pmMAR}}{Percent missing at random.}

\item{\code{extraOut}}{Additional outputs stored in a list.}

\item{\code{paramOnly}}{Logical indicating whether only parameters are stored.}

\item{\code{timing}}{Timing information for simulation runs.}
}}

\section{Objects from the Class}{

Objects can be created by \code{\link{sim}}.
}

