% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SimSem-class}
\alias{SimSem-class}
\alias{summary,SimSem-method}
\title{Class "SimSem"}
\description{
The template containing data-generation and data-analysis specification.
}
\section{Slots}{

\describe{
\item{\code{pt}}{Parameter table used in data analysis.}

\item{\code{dgen}}{Data generation template.}

\item{\code{modelType}}{Type of model (CFA, Path, or SEM) contained in this object.}

\item{\code{groupLab}}{The label of grouping variable.}

\item{\code{con}}{The list of defined parameters, equality constraints, or inequality
constraints specified in the model.}
}}

\section{Objects from the Class}{

Objects can be created by \code{\link{model}}.
}

