% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{simsem-package}
\alias{simsem}
\alias{simsem-package}
\title{simsem: SIMulated Structural Equation Modeling}
\description{
Provides an easy framework for Monte Carlo simulation in structural equation modeling, which can be used for various purposes, such as such as model fit evaluation, power analysis, or missing data handling and planning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://simsem.org/}
}

}
\author{
\strong{Maintainer}: Terrence D. Jorgensen \email{TJorgensen314@gmail.com} (\href{https://orcid.org/0000-0001-5111-6773}{ORCID})

Authors:
\itemize{
  \item Sunthud Pornprasertmanit \email{psunthud@gmail.com}
  \item Patrick Miller \email{pmille13@nd.edu}
  \item Alexander M. Schoemann \email{schoemanna@ecu.edu} (\href{https://orcid.org/0000-0002-8479-8798}{ORCID})
}

Other contributors:
\itemize{
  \item Corbin Quick \email{qcorbin@hsph.harvard.edu} [contributor]
}

}
\keyword{internal}
