% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SimVector-class}
\alias{SimVector-class}
\alias{summaryShort,SimVector-method}
\alias{summary,SimVector-method}
\title{Vector object: Random parameters vector}
\description{
This object can be used to represent a vector in an SEM model. It contains free
parameters, fixed values, starting values, and model misspecification.
This object can represent mean, intercept, or variance vectors.
}
\section{Slots}{

\describe{
\item{\code{free}}{The free-parameter vector. Any NA elements or character elements
are free. Any numeric elements are fixed as the specified number. If any free
elements have the same characters (except NA), the elements are equally constrained.}

\item{\code{popParam}}{Population parameter values.}

\item{\code{misspec}}{Model misspecification values.}
}}

\section{Objects from the Class}{

This object is created by \code{\link{bind}} function.
}

