% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPowerFitNonNested.R
\name{sortList}
\alias{sortList}
\title{Sort Two Objects Stored in a List}
\usage{
sortList(object)
}
\arguments{
\item{object}{A list containing two objects (e.g., vectors or matrices)
of equal dimensions.}
}
\value{
A list containing the sorted objects.
}
\description{
Swaps values between two objects in a list so that the first object
contains the smaller values and the second object contains the larger
values element-wise.
}
\keyword{internal}
