% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySimResult.R
\name{summaryConverge}
\alias{summaryConverge}
\title{Summarize convergence information}
\usage{
summaryConverge(object, std = FALSE, improper = TRUE)
}
\arguments{
\item{object}{A \code{SimResult} object.}

\item{std}{Logical indicating whether standardized population parameters
should be used.}

\item{improper}{Logical indicating whether improper solutions should be included.}
}
\value{
A list summarizing convergence statistics.
}
\description{
This function summarizes convergence results across simulation replications.
}
