% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySimResult.R
\name{summaryMisspec}
\alias{summaryMisspec}
\title{Summarize model misspecification}
\usage{
summaryMisspec(object, improper = TRUE)
}
\arguments{
\item{object}{A \code{SimResult} object.}

\item{improper}{Logical indicating whether improper solutions should be included.}
}
\value{
A data frame summarizing model misspecification and population fit values.
}
\description{
This function summarizes the obtained model misspecification values and
population fit indices.
}
