% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySimResult.R
\name{summaryPopulation}
\alias{summaryPopulation}
\title{Summarize population parameter values}
\usage{
summaryPopulation(object, std = FALSE, improper = TRUE)
}
\arguments{
\item{object}{A \code{SimResult} object.}

\item{std}{Logical indicating whether standardized population parameters
should be returned.}

\item{improper}{Logical indicating whether improper solutions should be included.}
}
\value{
A matrix summarizing the population parameter values.
}
\description{
This function summarizes the population parameter values used in the
data-generation model.
}
