% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySimResult.R
\name{summarySeed}
\alias{summarySeed}
\title{Summarize simulation seed}
\usage{
summarySeed(object)
}
\arguments{
\item{object}{A \code{SimResult} object.}
}
\value{
A list containing the seed information.
}
\description{
This function summarizes the seed values used during the simulation.
}
