% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryShort-methods.R
\name{summaryShort}
\alias{summaryShort}
\alias{summaryShort,ANY-method}
\title{Short summary of simulation objects}
\usage{
summaryShort(object, ...)

\S4method{summaryShort}{ANY}(object)
}
\arguments{
\item{object}{An object to summarize.}

\item{...}{Additional arguments passed to methods. For objects of class
\code{SimResult}, these include:
\describe{
  \item{alpha}{Significance level used when computing fit-index cutoffs
  for \code{SimResult} objects.}
  \item{digits}{Number of digits used when printing fit statistics.}
}}
}
\value{
This function prints a concise summary of the object. The returned value
is printed output describing the object structure, parameters, or results.
}
\description{
Provides a concise summary of objects used in the \code{simsem} framework.
If a specialized short summary is not available for an object class, the
function defaults to calling \code{\link{summary}}.
}
\details{
Methods are available for objects of class \code{SimMatrix}, \code{SimVector},
and \code{SimResult}.
}
\seealso{
\code{\link{summary}}, \code{\link{summaryFit}},
\code{\link{summaryPopulation}}, \code{\link{summaryMisspec}}
}
