% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySimResult.R
\name{summaryTime}
\alias{summaryTime}
\title{Summarize simulation timing}
\usage{
summaryTime(object, units = "seconds")
}
\arguments{
\item{object}{A \code{SimResult} object.}

\item{units}{Units used to display time.}
}
\value{
Printed timing summary.
}
\description{
This function summarizes the timing information recorded during the simulation.
}
