% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{temp.unstandardize.est.ov}
\alias{temp.unstandardize.est.ov}
\title{Internal helper for unstandardizing estimates}
\usage{
temp.unstandardize.est.ov(partable, ov.var = NULL, cov.std = TRUE)
}
\description{
Adapted from lavaan to avoid dependence on internal lavaan functions.
}
\keyword{internal}
