% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValueNonNested.R
\name{twoTailedPValue}
\alias{twoTailedPValue}
\title{Convert one-tailed p-values to two-tailed p-values}
\usage{
twoTailedPValue(vec)
}
\arguments{
\item{vec}{A vector of one-tailed \emph{p}-values.}
}
\value{
A vector of two-tailed \emph{p}-values.
}
\description{
Internal helper that converts one-tailed \emph{p}-values into
two-tailed \emph{p}-values.
}
\keyword{internal}
