% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrnorm.R
\name{vech2Corr}
\alias{vech2Corr}
\title{Convert vech to a Correlation Matrix}
\usage{
vech2Corr(vech)
}
\arguments{
\item{vech}{Vector of lower-triangular correlation values.}
}
\value{
A correlation matrix with ones on the diagonal.
}
\description{
Converts a vector containing the strictly lower triangular values
of a correlation matrix into a full symmetric correlation matrix.
}
\keyword{internal}
