\name{writeLavaanConstraint}
\alias{writeLavaanConstraint}
\title{
	Write a lavaan code for a given set of equality constraints
}
\description{
	Write a lavaan code for a given set of equality constraints
}
\usage{
writeLavaanConstraint(object, constraint)
}
\arguments{
  \item{object}{
	A parameter object, \code{\linkS4class{SimParam}}
}
  \item{constraint}{
	An equality constraint
}
}
\value{
	A \code{linkS4class{SimRSet} class with the text of lavaan constraint
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)
}
\examples{
# No example
}
