\name{writeLavaanNullCode}
\alias{writeLavaanNullCode}
\title{
	Write a lavaan code for a null model
}
\description{
	Write a lavaan code for a null model
}
\usage{
writeLavaanNullCode(var, aux = NULL)
}
\arguments{
  \item{var}{
	The name of variables in the model
}
  \item{aux}{
	The name of auxiliary variables
}
}
\value{
	A string containing the \code{lavaan} code for a null model
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)
}
\examples{
# No example
}
