\name{clean}
\alias{clean}
\title{
	Extract only converged replications in the result object
}
\description{
	Extract only converged replications in the result object (\code{\linkS4class{SimResult}})
}
\usage{
clean(object)
}
\arguments{
  \item{object}{
	The target result object (\code{\linkS4class{SimResult}})
}
}
\value{
	The cleaned result object
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; \email{psunthud@ku.edu})
}
\examples{
# No example
}
