\name{createData}
\alias{createData}
\title{
	TBA
}
\description{
	TBA
}
\usage{
createData(paramSet, n, indDist=NULL, sequential=FALSE, facDist=NULL, 
errorDist=NULL, indLab=NULL, modelBoot=FALSE, realData=NULL)
}
\arguments{
  \item{paramSet}{
	TBA
}
  \item{n}{
	Sample size
}
  \item{indDist}{
	TBA
}
  \item{sequential}{
	TBA
}
  \item{facDist}{
	TBA
}
  \item{errorDist}{
	TBA
}
  \item{indLab}{
	TBA
}
  \item{modelBoot}{
	TBA
}
  \item{realData}{
	TBA
}
}
\value{
	TBA
}
\author{
	Patrick Miller (Univeristy of Notre Dame; \email{pmille13@nd.edu})
    Sunthud Pornprasertmanit (University of Kansas; \email{psunthud@ku.edu})
}
\examples{
# No example
}
