\name{findphist}
\alias{findphist}
\title{
	Find the density (likelihood) of a pair value in 2D Kernel Density Estimate
}
\description{
	Find the density (likelihood) of a pair value in 2D Kernel Density Estimate
}
\usage{
findphist(value, hist) 
}
\arguments{
  \item{value}{
	A target pair of values
}
  \item{hist}{
	A 2D Binned Kernel Density Estimate 
}
}
\value{
	The probability (density) of the target pair of value
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; \email{psunthud@ku.edu})
}
\examples{
# No example
}
