\name{getKeywords}
\alias{getKeywords}
\title{
	List of all keywords used in the \code{simsem} package
}
\description{
	List of all keywords used in the \code{simsem} package
}
\usage{
getKeywords()
}
\value{
	A list of all keywords used in this package
	\itemize{
		\item \code{usedFit} Fit indices used as the default for providing output
		\item \code{usedFitPop} Population fit indices used as the default for providing input
		\item \code{optMin} The method picking the minimum value of misfit across misspecification sets
		\item \code{optMax} The method picking the maximum value of misfit across misspecification sets
		\item \code{optNone} Not using the optimization method
	}
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)
}
\examples{
# This function is not a public function.

# getKeywords()
}
