\name{VirtualDist-class}
\Rdversion{1.1}
\docType{class}
\alias{VirtualDist-class}
\title{Class "VirtualDist"}
\description{
All distribution objects. (Virtual Class)
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "VirtualDist" in the signature.
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; \email{psunthud@ku.edu})
}
\seealso{
	List of all distribution objects.
	\enumerate{
		\item \code{\linkS4class{SimNorm}}
		\item \code{\linkS4class{SimUnif}}
	}
}
\examples{
showClass("VirtualDist")
}
\keyword{classes}
