\name{simsem-package}
\alias{simsem-package}
\alias{simsem}
\docType{package}
\title{
SIMulated Structural Equation Modeling data. 
}
\description{
This package will generate data for structural equation modeling framework. This package is tailored to use those simulated data for various purposes, such as model fit evaluation.
}
\details{
\tabular{ll}{
Package: \tab simsem\cr
Type: \tab Package\cr
Version: \tab 0.0.7\cr
Depends: \tab R(>= 2.14), methods, lavaan, MASS\cr
Suggests: \tab parallel, Amelia, copula\cr
Date: \tab 2011-02-24\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)
Patrick Miller (University of Kansas; patr1ckm@ku.edu)
Alex Schoemann (University of Kansas; schoemann@ku.edu)
}
\keyword{ package }
\keyword{ sem }
\keyword{ simulation }
