\name{simsem-package}
\alias{simsem-package}
\alias{simsem}
\docType{package}
\title{
SIMulated Structural Equation Modeling. 
}
\description{
This package can be used to generate data using the structural equation modeling framework. This package is tailored to use those simulated data for various purposes, such as model fit evaluation, power analysis, or missing data handling and planning.
}
\details{
\tabular{ll}{
Package: \tab simsem\cr
Type: \tab Package\cr
Version: \tab 0.1.3\cr
Depends: \tab R(>= 2.14), methods, lavaan, MASS\cr
Suggests: \tab parallel, Amelia, copula, quantreg, splines, foreign\cr
Date: \tab 2012-05-06\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)
Patrick Miller (University of Kansas; patr1ckm@ku.edu)
Alex Schoemann (University of Kansas; schoemann@ku.edu)
}
\keyword{ package }
\keyword{ sem }
\keyword{ simulation }
