\name{simsem-package}
\alias{simsem-package}
\alias{simsem}
\docType{package}
\title{
SIMulated Structural Equation Modeling data. 
}
\description{
This package will generate data for structural equation modeling framework. This package is tailored to use those simulated data for various purposes, such as model fit evaluation.
}
\details{
\tabular{ll}{
Package: \tab simsem\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Depends: \tab R(>= 2.12), methods, lavaan, MASS\cr
Date: \tab 2011-09-22\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)

Maintainer: Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)
}
\keyword{ package }
\keyword{ sem }
\keyword{ simulation }
