\name{sortList}
\alias{sortList}
\title{
Sort two objects in a list 
}
\description{
Sort two objects in a list by swapping the values of both objects so that the first object contains the lower value and the second object contains the larger value
}
\usage{
sortList(object)
}
\arguments{
  \item{object}{
	The list with two objects (e.g., vector, matrix)
}
}
\value{
	The sorted list
}
\author{
    Sunthud Pornprasertmanit (University of Kansas; \email{psunthud@ku.edu})
}
\examples{
# No example
}
